fix(condo) INFRA-92: use only one context
